### 2022-01-27 #4102 @whedon
[PRE REVIEW]: DeepSynth: Scaling Neural Program Synthesis with Distribution-based Search
**Submitting author:** @Theomat (<a href="http://orcid.org/0000-0002-5043-3221">Théo Matricon</a>)
**Repository:** <a href="https://github.com/nathanael-fijalkow/DeepSynth" target ="_blank">https://github.com/nathanael-fijalkow/DeepSynth</a>
**Version:** v1.0
