+++ paper.bib
@@ -1,26 +1,0 @@
-@incollection{pytorch,
-title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
-author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
-booktitle = {Advances in Neural Information Processing Systems 32},
-editor = {H. Wallach and H. Larochelle and A. Beygelzimer and F. d\textquotesingle Alch\'{e}-Buc and E. Fox and R. Garnett},
-pages = {8024--8035},
-year = {2019},
-publisher = {Curran Associates, Inc.},
-url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf}
-}
-
-@misc{copilot,
-  author       = {GitHub},
-  title        = {GitHub Copilot},
-  url = {https://copilot.github.com/},
-  year         = {2021}
-}
-
-@misc{Fijalkow:2021,
-      title={Scaling Neural Program Synthesis with Distribution-based Search}, 
-      author={Nathanaël Fijalkow and Guillaume Lagarde and Théo Matricon and Kevin Ellis and Pierre Ohlmann and Akarsh Potta},
-      year={2021},
-      eprint={2110.12485},
-      archivePrefix={arXiv},
-      primaryClass={cs.LG}
-}
