### 2022-02-02 0bf49bf
Improved plotting
