### 2022-02-10 #4151 @whedon
[REVIEW]: DeepSynth: Scaling Neural Program Synthesis with Distribution-based Search
**Submitting author:** @Theomat (<a href="http://orcid.org/0000-0002-5043-3221">Théo Matricon</a>)
**Repository:** https://github.com/nathanael-fijalkow/DeepSynth
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0
