### 2022-02-11 #4151 @bzz
HI @ Theomat.

I’ve started reviewing the submission and got a couple of questions that are not immediately related to the code, so posting it here.

“What should my paper contain” of [JOSS  submission guideline](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) and “citations of relevant work” of[JOSS Review Criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#an-important-note-about-novel-software-and-citations-of-relevant-work) mentions

> list of key references, including to other software addressing related needs. Note that the references should include full names of venues, e.g., journals and conferences, not abbreviations only understood in the context of a specific discipline.

1. There is reference to Copilot, but no mention of the [Codex paper](https://arxiv.org/abs/2107.03374) that it is based on.
In this context, would you agree that the paper would benefit for a brief description on how this popular approach with large language models is fundamentally different from the neuro-symbolic one that your work is exploring?

2. I think the paper would also benefit greatly from more references to the other published OSS software in the field of program synthesis / programming by example (PbE)

Several examples that come to mind are:
  * [KasraF/SnipPy: Meta-repository for SnipPy](https://github.com/kasraF/Snippy)
  * [kensens/FrAngel: A Java program synthesizer.](https://github.com/kensens/FrAngel)
  * [utopia-group/regel: REGEL: Regular Expression Generation from Examples and Language](https://github.com/utopia-group/regel)
  * [StanfordPL/stoke: STOKE: A stochastic superoptimizer and program synthesizer](https://github.com/StanfordPL/stoke)

That may, in turn, call for including a short exposition on the enumerative/ stochastic/constraint-based search strategies, possibly with linking in a literature review like [Gulwani 2017](https://www.microsoft.com/en-us/research/wp-content/uploads/2017/10/program_synthesis_now.pdf) for further details.

I suggest all that, in the spirit of making whole program synthesis domain more legible for non-experts, but the curious souls with software engineering background who might be reading JOSS.

What do you think?

-Alex
