### 2022-02-14 6712100
Commented out test of no longer used function
