### 2022-02-14 cf85725
Updated test with new shapes
