### 2022-02-21 #4151 @Theomat
Hi Alex,
Thanks a lot for the suggestions!
1. About citing the Codex paper, we should have done it, thanks.
2. About giving a broader overview of techniques and OSS software in program synthesis, that's also a good point. However in 1000 words we won't be able to say much, so it's more about citing more relevant work? Another natural reference is the AAAI paper which uses the current tool for experiments; we did a more thorough job on related works -- we may simply refer to that section explicitly.
One last comment: many, if not most, of the program synthesis tools target a specific domain (DSL), and one (academic) appeal of DeepSynth is to work in a general fashion, allowing users to input DSLs themselves. For that reason, it's hard to list existing program synthesis tools, as each serves a different purpose.

- Nathanaël and Théo
