### 2022-02-26 nathanael-fijalkow/DeepSynth#2 @njuaplusplus
Questions about the installation instructions.
Hi Nathanaël,

I started to review your JOSS submission and have some questions about the installation part in the README.

### Quote the version specification
```
conda create -n deep_synth python>=3.7
```

should be changed to the following, otherwise, it's not accepted by some shells such as zsh.

```
conda create -n deep_synth "python>=3.7"
```

### How to install PyTorch
I would recommend providing the compatible PyTorch version requirements and some potential commands to install the compatible versions (such as different CUDA/CPU versions). Since conda env is already created, one can also install PyTorch via conda.

```
> pip install torch cython tqdm numpy matplotlib

ERROR: Could not find a version that satisfies the requirement torch (from versions: none)
ERROR: No matching distribution found for torch
```

### Missing pip package
```
pip install scipy  # required by unit_tests_algorithms.py
```

### Correct the script names

```
python unit_test_algorithms.py
python unit_test_programs.py
python unit_test_algorithms.py
python unit_test_predictions.py
# Only if you installed ray
python unit_test_parallel.py
```
The script name should be corrected.

```
python unit_tests_algorithms.py
python unit_tests_programs.py
python unit_tests_algorithms.py
python unit_tests_predictions.py
```

Missing file for unit_test_parallel.py.

### Fail to run the tests
```
> python unit_tests_algorithms.py
Traceback (most recent call last):
  File "/myapps/research/synthesis/DeepSynth/unit_tests_algorithms.py", line 11, in <module>
    from dsl import DSL
  File "/myapps/research/synthesis/DeepSynth/dsl.py", line 6, in <module>
    from cfg import CFG
  File "/myapps/research/synthesis/DeepSynth/cfg.py", line 4, in <module>
    from pcfg_logprob import LogProbPCFG
  File "/myapps/research/synthesis/DeepSynth/pcfg_logprob.py", line 7, in <module>
    import vose
  File "/home/aplusplus/anaconda3/envs/deep_synth/lib/python3.9/site-packages/vose/__init__.py", line 1, in <module>
    from .sampler import Sampler
  File "vose/sampler.pyx", line 1, in init vose.sampler
ValueError: numpy.ufunc size changed, may indicate binary incompatibility. Expected 232 from C header, got 216 from PyObject
```
A specific package version may be needed.

Best,
Shengwei
