### 2022-02-26 #4151 @njuaplusplus
Hi  @Theomat,

I have some problems completing the installation as instructed in your repo, could you take a look at this issue nathanael-fijalkow/DeepSynth#2 ? Thank you.

Best,
Shengwei
