### 2022-02-28 nathanael-fijalkow/DeepSynth#2 @nathanael-fijalkow
Hi Shengwei,

We have corrected what you pointed out, thanks a lot!
We also have added a requirement for the `pytorch` version (>=1.8) as we do
not need specific features.
The numpy error is due to a discrepancy between the used numpy version to
compile and the one used in the python environment, usually running `pip
upgrade numpy` fixes the issue even if `pip` tells you `numpy` is already
up to date. If this does not work, we have just tried the new order of
installation from scratch and haven't run into any issues.

Best,
Nath & Théo ***@***.***)
