### 2022-03-19 #4151 @njuaplusplus
Hi Nath & Théo,

Thank you very much for your patience. I have read your AAAI paper. It's quite interesting. The following are my suggestions on the documentation and software paper.

1. Example usage
I couldn't find detailed complete examples in the documentation. At least one complete example which can explore most of the features is expected.

2. Functionality documentation
No API method documentation.

3. Community guidelines
No such guidelines.

4. State of the field
This section mentions some packages but didn't describe the difference or comparison. Besides, I suggest citing and comparing [PROSE](https://github.com/microsoft/prose).

Best,
Shengwei
