### 2022-03-19 nathanael-fijalkow/DeepSynth#2 @njuaplusplus
Hi Nath & Théo,

There're some minor issues:

`pip upgrade numpy` => `pip install --upgrade numpy`

According to the files in your repo, `run_random_PCFG_search` => `run_randomPCFG_search`.

In order to run `run_random_PCFG_search.py`, one needs to `pip install matplotlib`. Please mention it in the documentation.

Thank you very much.

Best,
Shengwei
