### 2022-03-20 nathanael-fijalkow/DeepSynth#2 @Theomat
Hi Shengwei,

Thank you for the detailed corrections, they have been made.
For `matplotlib` although it is included in `conda install -c conda-forge cython tqdm numpy matplotlib scipy`, do you believe we need to mention it again?

Thanks

Best,
Théo
