+++ README.md
@@ -1,5 +1,5 @@
 # DeepSynth
-General-purpose program synthesiser.
+General-purpose program synthesizer.
 
 This is the repository for the code of the paper **"[Scaling Neural Program Synthesis with Distribution-based Search](https://arxiv.org/abs/2110.12485)"**
 published in the conference proceedings of the AAAI Conference on Artificial Intelligence, AAAI'22.
@@ -7,9 +7,27 @@ published in the conference proceedings of the AAAI Conference on Artificial Int
 **Authors**:
 Nathanaël Fijalkow, Guillaume Lagarde, Théo Matricon, Kevin Ellis, Pierre Ohlmann, Akarsh Potta
 
+<!-- toc -->
+
+- [Usage](#usage)
+  - [Installation](#installation)
+  - [File structure](#file-structure)
+- [Documentation and examples](#documentation-and-examples)
+    - [DSL](#dsl)
+    - [CFG and PCFG](#cfg-and-pcfg)
+    - [Synthesis](#synthesis)
+    - [Prediction from a model](#prediction-from-a-model)
+    - [Model creation](#model-creation)
+    - [Model training](#model-training)
+- [Reproducing the experiments](#reproducing-the-experiments)
+- [Report an issue](#report-an-issue)
+- [Contribute](#contribute)
+
+<!-- tocstop -->
+
 ![Figure](https://github.com/nathanael-fijalkow/DeepSynth/raw/main/main_figure.png)
 
-### Abstract
+## Abstract
 
  _We consider the problem of automatically constructing computer programs from input-output examples. We investigate
 how to augment probabilistic and neural program synthesis methods with new search algorithms, proposing a framework called distribution-based search. Within this framework,
@@ -68,7 +86,235 @@ python unit_tests_parallel.py
         Predictions/     # all files related to the ANN for prediction of the grammars 
 ```
 
-### Reproducing the experiments
+## Documentation and examples
+
+Table of contents:
+
+- [DSL](#dsl)
+- [CFG and PCFG](#cfg-and-pcfg)
+- [Synthesis](#synthesis)
+- [Prediction from a model](#prediction-from-a-model)
+- [Model creation](#model-creation)
+- [Model training](#model-training)
+
+### DSL
+
+For the dreamcoder dataset,, we have defined the DSL in ``DSL/deepcoder.py``.
+It contains two important objects: ``primitive_types`` and ``semantics``.
+The former is a dictionary mapping the primitives of the DSL to their types while the latter maps the primitives to their semantics, that is a value or function to be used when evaluating the primitive.
+Primitives can be constant such as 0, 1, 2... in the ``list.py`` DSL.
+If they are functions, then Python must be able to apply them one argument at a time, that would give the following semantic for ``+``: ``lambda a: lambda b: a + b``.
+
+There are two more subtleties in this file.
+First, the ``no_repetitions`` is a set of primitives than cannot be repeated such as ``SORT`` because it is useless to sort a sorted list, this reduces a bit the size of the program space.
+
+And the second are types. 
+Notice that we used ``Arrow``, ``List``, ``INT``... they come from ``from type_system import *``.
+There are ``PrimitiveType`` such as ``INT``, ``BOOL``, which are in truth nothing but a string, you can create your own with ``PrimitiveType(type_name)``.
+They have no constraints however DeepSynth makes the difference between them.
+Then there is ``List`` which makes up with ``List(type)`` a list of the given type.
[…]
