+++ paper.bib
@@ -1,3 +1,22 @@
+@article{Shi:2020,
+  author    = {Kensen Shi and
+               David Bieber and
+               Rishabh Singh},
+  title     = {TF-Coder: Program Synthesis for Tensor Manipulations},
+  journal   = {CoRR},
+  volume    = {abs/2003.09040},
+  year      = {2020},
+  url       = {https://arxiv.org/abs/2003.09040},
+}
+
+@InProceedings{Gulwani:2011,
+  author    = {Sumit Gulwani},
+  booktitle = {{ACM} {SIGPLAN-SIGACT} Symposium on Principles of Programming Languages, {POPL}},
+  title     = {Automating string processing in spreadsheets using input-output examples},
+  year      = {2011},
+  url       = {https://doi.org/10.1145/1926385.1926423},
+}
+
 @incollection{pytorch,
 title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
 author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
