+++ paper.bib
@@ -1,3 +1,37 @@
+@inproceedings{EllisWNSMHCST21,
+  author    = {Kevin Ellis and
+               Catherine Wong and
+               Maxwell I. Nye and
+               Mathias Sabl{\'{e}}{-}Meyer and
+               Lucas Morales and
+               Luke B. Hewitt and
+               Luc Cary and
+               Armando Solar{-}Lezama and
+               Joshua B. Tenenbaum},
+  title     = {DreamCoder: bootstrapping inductive program synthesis with wake-sleep library learning},
+  booktitle = {International Conference on Programming Language Design and Implementation, {PLDI}},
+  year      = {2021},
+  url       = {https://doi.org/10.1145/3453483.3454080},
+}
+
+@inproceedings{ZoharW18,
+  author    = {A. Zohar and L. Wolf},
+  title     = {Automatic Program Synthesis of Long Programs with a Learned Garbage Collector},
+  booktitle = {Neural Information Processing Systems, {NeurIPS}},
+  year      = {2018},
+  url       = {https://proceedings.neurips.cc/paper/2018/hash/390e982518a50e280d8e2b535462ec1f-Abstract.html},
+}
+
+@InProceedings{Balog2017,
+  author    = {Matej Balog and Alexander L. Gaunt and Marc Brockschmidt and Sebastian Nowozin and Daniel Tarlow},
+  booktitle = {International Conference on Learning Representations, {ICLR}},
+  title     = {DeepCoder: Learning to Write Programs},
+  year      = {2017},
+  bibsource = {dblp computer science bibliography, https://dblp.org},
+  biburl    = {https://dblp.org/rec/bib/conf/iclr/BalogGBNT17},
+  url       = {https://openreview.net/forum?id=ByldLrqlx},
+}
+
 @article{Shi:2020,
   author    = {Kensen Shi and
                David Bieber and
