### 2022-06-15 #4151 @Theomat
Hi,

We just added a paragraph about PROSE and rewrote the state of the art part in our paper.

Best,

Nath and Théo
