### 2022-07-10 f1ffa25
expanded README
