### 2022-07-19 #4151 @arfon
:wave: folks, as @terrytangyuan is stepping down from the JOSS editorial team (thanks for all of your efforts over the years Yuan!), I will be taking over this submission.

@njuaplusplus, @Kraks, @bzz – it looks like the authors have made a number of changes to the software and paper. Could I ask you to revisit this review and see if you're able to sign off on the remaining checklist items?

Many thanks!
