### 2022-08-28 #4151 @arfon
> @njuaplusplus, @Kraks, @bzz – it looks like the authors have made a number of changes to the software and paper. Could I ask you to revisit this review and see if you're able to sign off on the remaining checklist items?

Friendly reminder to check in again here soon please @njuaplusplus, @Kraks, and @bzz.
