+++ requirements.txt
@@ -1,0 +1,17 @@
+contourpy==1.0.5
+cycler==0.11.0
+Cython==0.29.32
+fonttools==4.37.4
+kiwisolver==1.4.4
+matplotlib==3.6.0
+numpy==1.23.3
+packaging==21.3
+Pillow==9.2.0
+pyparsing==3.0.9
+python-dateutil==2.8.2
+scipy==1.9.1
+six==1.16.0
+torch>=1.8
+tqdm==4.64.1
+typing_extensions==4.4.0
+vose @ git+https://github.com/MaxHalford/vose@a72ee6540c0d5b21158340d377ca5f11dc5ffb67
