### 2022-10-07 ed8af72
Doc: fix typo, remove duplicate commands
