+++ README.md
@@ -73,7 +73,6 @@ pip install --upgrade numpy
 # To test your installation you can run the following tests:
 python unit_tests_algorithms.py
 python unit_tests_programs.py
-python unit_tests_algorithms.py
 python unit_tests_predictions.py
 # Only if you installed ray
 python unit_tests_parallel.py
@@ -102,7 +101,7 @@ Table of contents:
 
 ### DSL
 
-For the dreamcoder dataset,, we have defined the DSL in ``DSL/deepcoder.py``.
+For the dreamcoder dataset, we have defined the DSL in ``DSL/deepcoder.py``.
 It contains two important objects: ``primitive_types`` and ``semantics``.
 The former is a dictionary mapping the primitives of the DSL to their types while the latter maps the primitives to their semantics, that is a value or function to be used when evaluating the primitive.
 Primitives can be constant such as 0, 1, 2... in the ``list.py`` DSL.
