### 2022-10-07 #4151 @bzz
Tried out the software package, check more boxes and sent some minor suggestions at https://github.com/nathanael-fijalkow/DeepSynth/pull/3
