### 2022-10-10 #4151 @Theomat
Hi bzz,

Thank you for your pull request!

Defining a DSL is highly dependent on the domain you're targeting, the more exotic from the current ones the larger the work. There is no single way to do it nor a manual for that. E.g.: let's say I want a DSL to draw functions then I need to define all the language elements then add something to evaluate those new programs then add a way to train such a new model then I would need to generate a training dataset. This is a rather complicated issue that is left to the user; unfortunately there is no generic way to do that. We have worked out some examples, but each new one requires some thinking!

Théo and Nath
