### 2022-10-13 #4151 @arfon
@Theomat – At this point, with two complete reviews, we're ready to move forward. Could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
