### 2022-10-16 #4151 @arfon
@Theomat – can you check if the DOIs that @editorialbot is suggesting are correct for your paper and if they are, could you add them to your BibTeX file please?
