### 2022-10-16 #4151 @arfon
@Theomat – I've also found a few small issues with your paper that need fixing: https://github.com/nathanael-fijalkow/DeepSynth/pull/4
