+++ paper.bib
@@ -12,6 +12,7 @@
   booktitle = {International Conference on Programming Language Design and Implementation, {PLDI}},
   year      = {2021},
   url       = {https://doi.org/10.1145/3453483.3454080},
+  doi       = {10.1145/3453483.3454080},
 }
 
 @inproceedings{ZoharW18,
@@ -41,6 +42,7 @@
   volume    = {abs/2003.09040},
   year      = {2020},
   url       = {https://arxiv.org/abs/2003.09040},
+  doi       = {10.1145/3517034},
 }
 
 @InProceedings{Gulwani:2011,
@@ -49,6 +51,8 @@
   title     = {Automating string processing in spreadsheets using input-output examples},
   year      = {2011},
   url       = {https://doi.org/10.1145/1926385.1926423},
+  doi       = {10.1145/1926385.1926423},
+  
 }
 
 @incollection{pytorch,
@@ -69,11 +73,5 @@ url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-perf
   year         = {2021}
 }
 
-@misc{Fijalkow:2021,
-      title={Scaling Neural Program Synthesis with Distribution-based Search}, 
-      author={Nathanaël Fijalkow and Guillaume Lagarde and Théo Matricon and Kevin Ellis and Pierre Ohlmann and Akarsh Potta},
-      year={2021},
-      eprint={2110.12485},
-      archivePrefix={arXiv},
-      primaryClass={cs.LG}
-}
+
+@article{Fijalkow:2021, title={Scaling Neural Program Synthesis with Distribution-Based Search}, volume={36}, url={https://ojs.aaai.org/index.php/AAAI/article/view/20616}, DOI={10.1609/aaai.v36i6.20616}, number={6}, journal={Proceedings of the AAAI Conference on Artificial Intelligence}, author={Fijalkow, Nathanaël and Lagarde, Guillaume and Matricon, Théo and Ellis, Kevin and Ohlmann, Pierre and Potta, Akarsh Nayan}, year={2022}, month={Jun.}, pages={6623-6630} }
