### 2022-10-16 #4151 @Theomat
Hi,

All the missing DOIs pointed out by the bot have been added.
